[post]でのTextをコンポーネントで作成
